Clean server side

+++ Client-server/src/main/resources/db.migration/V1__Init_DB.sql
@@ -1,14 +1,10 @@
 CREATE TABLE "Object" (
      "id" serial NOT NULL,
-     "lead_id" serial NOT NULL,
+     "lead_id" bigint NOT NULL,
      "lat" float8 NOT NULL,
      "lng" float8 NOT NULL,
      CONSTRAINT "Object_pk" PRIMARY KEY ("id")
 );
-
-
-
-
